refactor: 기존 어드민 api에서 도메인명으로 명확한 역할 분리
